requirements.txt
pyyaml
jupyter
chardet
requests
pytest
pytest-cov
fabric3
